improve helpers separation and update existing tests
