+++ tests/e2e/tests/add-mongodb-datasource-flow.spec.ts
@@ -1,17 +1,17 @@
 import { type Page, test, expect } from '@playwright/test';
-import { performInitialSetup, navigateToSettings, navigateToDataSourceForm } from '../helpers/setup';
+import { performInitialSetup } from '../helpers/setup';
+import { navigateToDataSourceForm, navigateToSettings } from '../helpers/navigate';
 
 test.describe('Add MongoDB Data Source Flow', () => {
   test.beforeEach(async ({ page }) => {
     test.setTimeout(120000);
     await performInitialSetup(page);
     await navigateToSettings(page);
+    await navigateToDataSourceForm(page, 'mongodb');
   });
 
   test('Create MongoDB data source with valid connection string', async ({ page }: { page: Page }) => {
     console.log('Starting MongoDB data source creation test...');
-
-    await navigateToDataSourceForm(page, 'mongodb');
 
     console.log('🔍 Looking for database name input...');
 
@@ -70,8 +70,6 @@ test.describe('Add MongoDB Data Source Flow', () => {
 
   test('Validate MongoDB connection string format and required fields', async ({ page }: { page: Page }) => {
     console.log('Starting MongoDB validation test...');
-
-    await navigateToDataSourceForm(page, 'mongodb');
 
     // Test 1: Try to create without filling required fields
     console.log('🔍 Testing form validation - attempting to create without required fields...');
